# openapi/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "poc-quarkus API",
        "version": "0.0.1"
    },
    "paths": {
        "/info": {
            "get": {
                "tags": [
                    "Home Controller"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AppInfo"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/organizations": {
            "get": {
                "tags": [
                    "Enrollments Controller"
                ],
                "responses": {
                    "200": {
                        "description": "Obtained all enrollments.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrganizationModelResponse"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Service unavailable.",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/organizations/{organizationFiscalCode}": {
            "get": {
                "tags": [
                    "Enrollments Controller"
                ],
                "parameters": [
                    {
                        "name": "organizationFiscalCode",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obtained single enrollment.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrganizationModelResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not found the enroll service.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Service unavailable.",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Enrollments Controller"
                ],
                "parameters": [
                    {
                        "name": "organizationFiscalCode",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrganizationModelResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthourized",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Enrollments Controller"
                ],
                "parameters": [
                    {
                        "name": "organizationFiscalCode",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request deleted.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrganizationModelResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthourized.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not found the creditor institution.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Service unavailable.",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AppInfo": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "version": {
                        "type": "string"
                    },
                    "environment": {
                        "type": "string"
                    }
                }
            },
            "LocalDateTime": {
                "format": "date-time",
                "type": "string",
                "example": "2022-03-10T12:15:50"
            },
            "OrganizationModelResponse": {
                "type": "object",
                "properties": {
                    "organizationFiscalCode": {
                        "type": "string"
                    },
                    "organizationOnboardingDate": {
                        "$ref": "#/components/schemas/LocalDateTime"
                    }
                }
            }
        }
    }
}